Rename clients & optimize selection
